Refactor timer callback to 'onHold' from 'onPress'
Previosly, the callback could easily get called and activated
because of a miss-press. Also, the timming on multi-presses was
quite tight, to avoid miss-presses.

Making the callback depend on holding the button for sometime
(currently, 500ms) fixes thoses issues two issues.

The delay introduced by checking the button for some time is
removed from the accumulated time. This does cause the timer to
move 500ms back when unpaused.